--- UCLL.projects.WeatherStations.sln
@@ -1,53 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35312.102
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionItems", "{4A6C06FF-EA7D-4406-9BC4-859684C1F0FF}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{2215AC95-1A8B-4183-95DE-01908FF7823F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{EB03D4F0-E950-412B-967A-5C21D9E03114}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docker", "docker", "{6F77E1AD-947F-4C5D-98F0-3ACED2D7C14E}"
-	ProjectSection(SolutionItems) = preProject
-		docker\compose.yml = docker\compose.yml
-		docker\.env = docker\.env
-		docker\.env.sample = docker\.env.sample
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "docker-data", "docker-data", "{2725C35D-EEC9-4E51-98BD-E791C7234828}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UCLL.projects.WeatherStations.WebApi", "src\UCLL.projects.WeatherStations.WebApi\UCLL.projects.WeatherStations.WebApi.csproj", "{C057591D-6F9A-4E08-B687-CD301892F213}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UCLL.projects.WeatherStations.MQTT", "src\UCLL.projects.WeatherStations.MQTT\UCLL.projects.WeatherStations.MQTT.csproj", "{338E1417-ADC3-4E83-8597-2A075C025BE9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C057591D-6F9A-4E08-B687-CD301892F213}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C057591D-6F9A-4E08-B687-CD301892F213}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C057591D-6F9A-4E08-B687-CD301892F213}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C057591D-6F9A-4E08-B687-CD301892F213}.Release|Any CPU.Build.0 = Release|Any CPU
-		{338E1417-ADC3-4E83-8597-2A075C025BE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{338E1417-ADC3-4E83-8597-2A075C025BE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{338E1417-ADC3-4E83-8597-2A075C025BE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{338E1417-ADC3-4E83-8597-2A075C025BE9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B8A7FA4D-A054-4660-82BF-71BEB15D2E1F}
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{C057591D-6F9A-4E08-B687-CD301892F213} = {2215AC95-1A8B-4183-95DE-01908FF7823F}
-		{338E1417-ADC3-4E83-8597-2A075C025BE9} = {2215AC95-1A8B-4183-95DE-01908FF7823F}
-	EndGlobalSection
-EndGlobal
