+++ src/UCLL.Projects.WeatherStations.WebApi/Dto/MeasurementDto.cs
@@ -1,0 +1,9 @@
+﻿namespace UCLL.Projects.WeatherStations.WebApi.Dto
+{
+    public class MeasurementDto
+    {
+        public DateTime Timestamp { get; set; }
+        public double Value { get; set; }
+    }
+
+}
