+++ src/UCLL.Projects.WeatherStations.WebApi/Interfaces/IStationRepository.cs
@@ -1,0 +1,15 @@
+﻿#region
+
+using UCLL.Projects.WeatherStations.WebApi.Dto;
+
+#endregion
+
+namespace UCLL.Projects.WeatherStations.WebApi.Interfaces
+{
+    public interface IStationRepository
+    {
+        //List<Station> GetAllStations();
+
+        public IEnumerable<StationDto> GetStationsLatestMeasurements(List<int> stationIds, int measurementAmount);
+    }
+}
