+++ src/UCLL.Projects.WeatherStations.WebApi/Models/Measurement.cs
@@ -1,0 +1,13 @@
+﻿namespace UCLL.Projects.WeatherStations.WebApi.Models
+{
+    public class Measurement
+    {
+        public DateTime Timestamp { get; set; }
+
+        public int StationId { get; set; }
+
+        public int SensorId { get; set; }
+
+        public double Value { get; set; }
+    }
+}
