+++ src/UCLL.Projects.WeatherStations.WebApi/Properties/launchSettings.json
