+++ src/UCLL.Projects.WeatherStations.WebApi/WeatherStationAPI.http
