+++ src/UCLL.Projects.WeatherStations.WebApi/appsettings.json
