--- src/UCLL.projects.WeatherStations.MQTT/Dockerfile
@@ -1,21 +1,0 @@
-﻿FROM mcr.microsoft.com/dotnet/runtime:8.0 AS base
-USER $APP_UID
-WORKDIR /app
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["UCLL.projects.WeatherStations.MQTT/UCLL.projects.WeatherStations.MQTT.csproj", "UCLL.projects.WeatherStations.MQTT/"]
-RUN dotnet restore "UCLL.projects.WeatherStations.MQTT/UCLL.projects.WeatherStations.MQTT.csproj"
-COPY . .
-WORKDIR "/src/UCLL.projects.WeatherStations.MQTT"
-RUN dotnet build "UCLL.projects.WeatherStations.MQTT.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "UCLL.projects.WeatherStations.MQTT.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "UCLL.projects.WeatherStations.MQTT.dll"]
