--- src/UCLL.projects.WeatherStations.WebApi/Dockerfile
@@ -1,24 +1,0 @@
-﻿FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-USER $APP_UID
-WORKDIR /app
-
-EXPOSE 8080
-EXPOSE 8081
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["UCLL.projects.WeatherStations.WebApi/UCLL.projects.WeatherStations.WebApi.csproj", "UCLL.projects.WeatherStations.WebApi/"]
-RUN dotnet restore "UCLL.projects.WeatherStations.WebApi/UCLL.projects.WeatherStations.WebApi.csproj"
-COPY . .
-WORKDIR "/src/UCLL.projects.WeatherStations.WebApi"
-RUN dotnet build "UCLL.projects.WeatherStations.WebApi.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "UCLL.projects.WeatherStations.WebApi.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "UCLL.projects.WeatherStations.WebApi.dll"]
