--- src/UCLL.projects.WeatherStations.WebApi/Dto/MeasurementDto.cs
@@ -1,9 +1,0 @@
-﻿namespace UCLL.projects.WeatherStations.WebApi.Dto
-{
-    public class MeasurementDto
-    {
-        public DateTime Timestamp { get; set; }
-        public double Value { get; set; }
-    }
-
-}
