--- src/UCLL.projects.WeatherStations.WebApi/Dto/SensorDto.cs
@@ -1,11 +1,0 @@
-﻿namespace UCLL.projects.WeatherStations.WebApi.Dto
-{
-    public class SensorDto
-    {
-        public int Id { get; set; }
-        public string Unit { get; set; }
-        public string Type { get; set; }
-
-        public List<MeasurementDto>? Measurements { get; set; }
-    }
-}
