--- src/UCLL.projects.WeatherStations.WebApi/Models/Measurement.cs
@@ -1,13 +1,0 @@
-﻿namespace UCLL.projects.WeatherStations.WebApi.Models
-{
-    public class Measurement
-    {
-        public DateTime Timestamp { get; set; }
-
-        public int StationId { get; set; }
-
-        public int SensorId { get; set; }
-
-        public double Value { get; set; }
-    }
-}
