--- src/UCLL.projects.WeatherStations.WebApi/Models/Station_Sensor.cs
@@ -1,13 +1,0 @@
-﻿namespace UCLL.projects.WeatherStations.WebApi.Models
-{
-    public class Station_Sensor
-    {
-        public int StationId { get; set; }
-
-        public int SensorId { get; set; }
-
-        //relationships
-        public Station Station { get; set; } = null!;
-        public Sensor Sensor { get; set; } = null!;
-    }
-}
